api/permissions.py
from rest_framework.permissions import BasePermission


class IsAuthorOrReadOnlyPermission(BasePermission):

    def has_object_permission(self, request, view, obj):
        if request.method in ('DELETE', 'PUT', 'PATCH'):
            return request.user == obj.author
        return True
