requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
pillow
sorl-thumbnail
markdown
django-cors-headers
django-filters